Fix audio creation logic
